+++ src/winenv/shells.py
@@ -1,0 +1,37 @@
+import abc
+
+
+class ShellSyntax(abc.ABC):
+
+    @abc.abstractmethod
+    def export_variable(self, variable, value):
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def unset_variable(self, variable):
+        raise NotImplementedError
+
+
+class BashSyntax(ShellSyntax):
+
+    def export_variable(self, variable, value):
+        return 'export {}="{}"'.format(variable, value)
+
+    def unset_variable(self, variable):
+        return 'unset {}'.format(variable)
+
+
+class FishSyntax(ShellSyntax):
+
+    def export_variable(self, variable, value):
+        return 'set -x {} "{}"'.format(variable, value)
+
+    def unset_variable(self, variable):
+        return 'set -e {}'.format(variable)
+
+SHELLS = {
+    'bash': BashSyntax(),
+    'fish': FishSyntax(),
+}
+
+DEFAULT_SHELL = 'bash'