+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project uses [semantic versioning](http://semver.org/).
 ### Added
 
 - Added list command.
+- Added support for fish and shell selection.
 
 ## 0.1.0